pyproject.toml
[project]
name = "parachute-computer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0.1",
    "claude-agent-sdk>=0.1.29",
    "sse-starlette>=2.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.scripts]
parachute = "parachute.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
